rename files again
