+++ src/types/notificationEvents.ts
